summit_seo/web/api/requirements.txt
# API Dependencies
fastapi>=0.95.0,<0.96.0
uvicorn>=0.21.1,<0.22.0
python-dotenv>=1.0.0,<1.1.0
sqlalchemy>=2.0.7,<2.1.0
pydantic>=1.10.7,<1.11.0
alembic>=1.10.3,<1.11.0
httpx>=0.23.3,<0.24.0
pytest>=7.3.0,<7.4.0
pytest-asyncio>=0.21.0,<0.22.0
python-multipart>=0.0.6,<0.1.0
bcrypt>=4.0.1,<4.1.0
pyjwt>=2.6.0,<2.7.0
passlib>=1.7.4,<1.8.0
email-validator>=2.0.0,<2.1.0

# Supabase Integration
supabase>=1.0.3,<1.1.0
psycopg2-binary>=2.9.6,<2.10.0  # For PostgreSQL connection
python-jose>=3.3.0,<3.4.0  # For JWT handling

# System Monitoring
psutil>=5.9.0,<5.10.0

# API Documentation
pyyaml>=6.0,<6.1.0
